+++ app/(auth)/otp-verification/page.tsx
@@ -41,7 +41,7 @@ export default function VerifyRegistrationOTP() {
 
   return (
     <div className="flex min-h-screen">
-      <div className="w-full md:w-2/3 flex flex-col py-10 bg-white">
+      <div className="w-full md:w-2/3 lg:w-1/2 flex flex-col py-10 bg-white">
         <div className="px-3 md:px-5 mb-15">
           <Image
             aria-hidden
@@ -94,13 +94,16 @@ export default function VerifyRegistrationOTP() {
         </div>
       </div>
 
-      <div className="hidden md:block w-[50%] relative overflow-hidden bg-gradient-to-t from-[#C7DCD2] via-[#E5EFEA] to-[#F9FBFA]">
-        <Image
-          src="/wastecan1.svg"
-          alt="Waste Bin"
-          fill
-          className="object-contain object-right bg-inherit"
-        />
+      <div className="relative hidden md:flex md:w-1/3 lg:w-1/2  overflow-hidden bg-gradient-to-t from-[#C7DCD2] via-[#E5EFEA] to-[#F9FBFA]">
+        <div className="absolute bottom-0 md:right-0 xl:left-1/2 xl:-translate-x-1/2">
+          <Image
+            src="/wastecan.png"
+            alt="waste-can"
+            width={400}
+            height={400}
+            className="object-contain md:w-[280px] lg:w-[350px] xl:w-[500px]"
+          />
+        </div>
       </div>
     </div>
   );
