+++ app/(auth)/register/page.tsx
@@ -95,7 +95,7 @@ export default function SignupForm() {
   return (
     <div className="flex min-h-screen">
       {/* flex form and bg image */}
-      <div className="w-full md:w-2/3 flex flex-col py-10 bg-white">
+      <div className="w-full md:w-2/3 lg:w-1/2 flex flex-col py-10 bg-white">
         {/* Logo */}
         <Link href="/" className="px-3 md:px-5 mb-15">
           <Image
@@ -615,12 +615,10 @@ export default function SignupForm() {
           </div>
         </div>
       </div>
-      <div className="hidden md:block w-[50%] relative overflow-hidden bg-gradient-to-t from-[#C7DCD2] via-[#E5EFEA] to-[#F9FBFA]">
-        <div className="w-full max-w-[300px] px-10 mt-20 relative z-10">
+      <div className="relative hidden md:flex md:w-1/3 lg:w-1/2  overflow-hidden bg-gradient-to-t from-[#C7DCD2] via-[#E5EFEA] to-[#F9FBFA]">
+        <div className="w-full md:max-w-[300px] lg:max-w-[400px] px-10 mt-20 relative z-10">
           <h3 className="text-primary text-xl font-bold">
-            Register an
-            <br />
-            Account
+            Register an Account
           </h3>
           <p className="text-primary mt-3">
             Register an account as an individual or business to access all the
@@ -630,21 +628,39 @@ export default function SignupForm() {
         </div>
 
         {option === "individual" && (
-          <Image
-            src="/wastecan1.svg"
-            alt="waste-can"
-            fill
-            className="object-contain object-right z-0"
-          />
+          <div className="absolute bottom-0 md:right-0 xl:left-1/2 xl:-translate-x-1/2">
+            <Image
+              src="/wastecan.png"
+              alt="waste-can"
+              width={400}
+              height={400}
+              className="object-contain md:w-[280px] lg:w-[350px] xl:w-[500px]"
+            />
+          </div>
         )}
 
         {option === "business" && (
-          <Image
-            src="/wastecan2.svg"
-            alt="Waste Bin"
-            fill
-            className="object-contain object-right z-0"
-          />
+          <>
+            <div className="absolute bottom-0 md:right-0">
+              <Image
+                src="/truck.png"
+                alt="Waste Bin"
+                width={400}
+                height={400}
+                className="object-contain md:w-[300px] lg:w-[400px] xl:w-[550px]"
+              />
+            </div>
+
+            <div className="absolute bottom-0  right-[20%] md:right-[25%] lg:right-[28%] xl:right-[30%]">
+              <Image
+                src="/trash.png"
+                alt="Waste Bin"
+                width={400}
+                height={400}
+                className="object-contain md:w-[80px] xl:w-[100px]"
+              />
+            </div>
+          </>
         )}
       </div>
     </div>
