+++ package.json
@@ -25,6 +25,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
+    "iconsax-react": "^0.0.8",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.536.0",
     "next": "15.4.5",
